ref: Move tracing utils to separate file and log if no routing inst
