+++ src/js/tracing/utils.ts
@@ -1,0 +1,34 @@
+import { IdleTransaction, SpanStatus } from "@sentry/tracing";
+
+/**
+ * Converts from milliseconds to seconds
+ * @param time time in ms
+ */
+export function msToSec(time: number): number {
+  return time / 1000;
+}
+
+/**
+ * Converts from seconds to milliseconds
+ * @param time time in seconds
+ */
+export function secToMs(time: number): number {
+  return time * 1000;
+}
+
+/**
+ *
+ */
+export function adjustTransactionDuration(
+  maxDuration: number,
+  transaction: IdleTransaction,
+  endTimestamp: number
+): void {
+  const diff = endTimestamp - transaction.startTimestamp;
+  const isOutdatedTransaction =
+    endTimestamp && (diff > maxDuration || diff < 0);
+  if (isOutdatedTransaction) {
+    transaction.setStatus(SpanStatus.DeadlineExceeded);
+    transaction.setTag("maxTransactionDurationExceeded", "true");
+  }
+}
